% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaguess.R
\name{betaguess}
\alias{betaguess}
\title{Guess the beta-distribution parameters based on CI}
\usage{
betaguess(best, lower, upper, p = 0.95, method = "mode")
}
\arguments{
\item{best}{The highest probability (see \code{method}).}

\item{lower}{The lower level of the \code{p} confidence interval.}

\item{upper}{The upper level of the \code{p} confidence interval.}

\item{p}{The confidence interval (\code{default}: 95\%).}

\item{method}{Choose mode or mean as best estimate.}
}
\value{
The \code{alpha} and \code{beta} parameters of the beta-distributions.
}
\description{
Guess the beta-distribution parameters based on CI
}
\examples{
\dontrun{
betaguess(best, lower, upper, p = 0.95, method = "mode")
}
}
