% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dtaportal.R
\name{check_dtaportal}
\alias{check_dtaportal}
\title{check and struturize dataportal file}
\usage{
check_dtaportal(x)
}
\arguments{
\item{x}{A data.frame or tibble}
}
\value{
data.frame
}
\description{
check and struturize dataportal file
}
