% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{diff_perc}
\alias{diff_perc}
\title{Calculate difference in percentages (\%)}
\usage{
diff_perc(x, y, dec = 1, as_text = TRUE)
}
\arguments{
\item{x}{The reference proportion.}

\item{y}{The most recent value upon which the difference should be calculated.}

\item{dec}{The number of decinimal numbers to be used.}

\item{as_text}{Should the result be transformed as text.}
}
\value{
difference (in \%).
}
\description{
Calculate difference in percentages (\%)
}
