% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharter.R
\name{drop_padding}
\alias{drop_padding}
\title{Highcharter fx: drop padding}
\usage{
drop_padding(x)
}
\arguments{
\item{x}{A highcharter chart}
}
\value{
A highchater chart.
}
\description{
Highcharter fx: drop padding
}
