% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharter.R
\name{drop_xAxis}
\alias{drop_xAxis}
\title{Highcharter fx: drop xAxis}
\usage{
drop_xAxis(x)
}
\arguments{
\item{x}{A highcharter chart}
}
\value{
A highchater chart.
}
\description{
Highcharter fx: drop xAxis
}
