% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharter.R
\name{drop_yAxis}
\alias{drop_yAxis}
\title{Highcharter fx: drop yAxis}
\usage{
drop_yAxis(x)
}
\arguments{
\item{x}{A highcharter chart}
}
\value{
A highchater chart.
}
\description{
Highcharter fx: drop yAxis
}
