% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dw_calc.R
\name{dw_calc}
\alias{dw_calc}
\title{Perform the standard DW calculations (non-stratified)}
\usage{
dw_calc(xlsx, dis)
}
\arguments{
\item{xlsx}{Excel-file with sheets named as in input dis.}

\item{dis}{names of the variables that should be extracted to calculate the disease DW (!should be character string).}
}
\value{
List with resulting DW age- and gender stratified for each of the given diseases.
}
\description{
Perform the standard DW calculations (non-stratified)
}
