% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharter.R
\name{export_hc2}
\alias{export_hc2}
\title{Highcharter fx: export}
\usage{
export_hc2(hc, filename, map = FALSE)
}
\arguments{
\item{hc}{A highcharter chart}

\item{filename}{Filename used for saving}

\item{map}{Are you exporting a map? (Default: \code{FALSE})}
}
\value{
Saved highcharter chart in .js
}
\description{
Highcharter fx: export
}
