% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getage}
\alias{getage}
\title{Recompile given age data to R-compatible age groups}
\usage{
getage(x)
}
\arguments{
\item{x}{Age groups as character string. It is important that age categories start
with < X, and end with X+. In-between categories should be defined as X-X.}
}
\value{
Age group as R-compatible factors.
}
\description{
Recompile given age data to R-compatible age groups
}
