% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getage_num}
\alias{getage_num}
\title{Recompile continuous age data to R-compatible age groups}
\usage{
getage_num(x)
}
\arguments{
\item{x}{Age groups as continuous variable.}
}
\value{
Age group as R-compatible factors.
}
\description{
Recompile continuous age data to R-compatible age groups
}
