% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YLD.R
\name{getYLD}
\alias{getYLD}
\title{Calculate YLD}
\usage{
getYLD(x, f_prev, f_dw, POP, folder)
}
\arguments{
\item{x}{Short code of the disease for which YLD should be calculated}

\item{f_prev}{File with prevalence estimates}

\item{f_dw}{File with calculated DW (disease model)}

\item{POP}{Population datafile from STATBEL (cleaned)}

\item{folder}{Folder to save output}
}
\value{
Saved csv-file
}
\description{
Calculate YLD
}
