% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/his_helpers.R
\name{his_management}
\alias{his_management}
\title{Perform the datamanagement of the HIS dataset for the BeBod-project}
\usage{
his_management(df_his, dis, year = "all")
}
\arguments{
\item{df_his}{raw BEHIS dataset that can be read in by the haven-package.}

\item{dis}{names of the variables that should be extracted to calculate the disease prevalences (!should be character string)}

\item{year}{years that should be extracted from the HIS dataset. Default value: "all" for all years.}
}
\value{
Clean HIS dataset
}
\description{
Perform the datamanagement of the HIS dataset for the BeBod-project
}
