% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/his_helpers.R
\name{his_prev_calc}
\alias{his_prev_calc}
\title{Perform prevalence calculations for the HIS dataset}
\usage{
his_prev_calc(df_his, HIS, ref_year)
}
\arguments{
\item{df_his}{cleaned BEHIS dataset that resulted from the cleaning in the his_management-function.}

\item{HIS}{dataframe containing the following columns: ID - ID number, DIS - abbr. disease, HIS - name of DIS in the HIS dataset, and SUBJ - should subjective health be taken into account.}

\item{ref_year}{Reference year for the calculation based on the HIS.}
}
\value{
list of prevalence estimates for each disease that is included in the HIS dataset (NA will be discarted).
}
\description{
Perform prevalence calculations for the HIS dataset
}
