% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ima_helpers.R
\name{ima_management}
\alias{ima_management}
\title{Perform the datamanagement of the IMA dataset for the BeBod-project}
\usage{
ima_management(df_ima, dis, residence = "resident")
}
\arguments{
\item{dis}{names of the variables that should be extracted to calculate the disease prevalences (!should be character string)}

\item{residence}{residence status that should be extracted from the IMA dataset. Default value: "resident" for Belgian residents.}

\item{df_his}{raw IMA dataset that can be read in by the haven-package.}
}
\value{
Clean IMA dataset
}
\description{
Perform the datamanagement of the IMA dataset for the BeBod-project
}
