% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ima_helpers.R
\name{ima_prev_calc}
\alias{ima_prev_calc}
\title{Perform prevalence calculations for the HIS dataset}
\usage{
ima_prev_calc(df_ima, IMA)
}
\arguments{
\item{df_ima}{cleaned BEHIS dataset that resulted from the cleaning in the his_management-function.}

\item{IMA}{dataframe containing the following columns: ID - ID number, DIS - abbr. disease, HIS - name of DIS in the HIS dataset, and SUBJ - should subjective health be taken into account.}
}
\value{
list of prevalence estimates for each disease that is included in the HIS dataset (NA will be discarted).
}
\description{
Perform prevalence calculations for the HIS dataset
}
