% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intego_helpers.R
\name{intego_management}
\alias{intego_management}
\title{Perform the datamanagement of the INTEGO dataset for the BeBod-project}
\usage{
intego_management(df_intego)
}
\arguments{
\item{df_intego}{raw INTEGO dataset that can be read in by the open.xlsx package..}
}
\value{
INTEGO dataset
}
\description{
Perform the datamanagement of the INTEGO dataset for the BeBod-project
}
