% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_ci.R
\name{mean_ci}
\alias{mean_ci}
\title{Calculate mean and 95\% CI of distribution}
\usage{
mean_ci(x)
}
\arguments{
\item{x}{A continuous variable.}
}
\value{
The mean and 95\% CI of \code{x}.
}
\description{
Calculate mean and 95\% CI of distribution
}
\examples{
\dontrun{
mean_ci(x)
}
}
