% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{midyear}
\alias{midyear}
\title{Calculate midyear population}
\usage{
midyear(x, year_ref, year_plus_one)
}
\arguments{
\item{x}{Statbel DF with variable "YEAR".}

\item{year_ref}{Reference year for which to calculate mid-year population}

\item{year_plus_one}{Reference year + 1 (!should be available in the dataset)}
}
\value{
mid-population dataframe.
}
\description{
Calculate midyear population
}
