% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pop_distribution}
\alias{pop_distribution}
\title{Calculate population distribution}
\usage{
pop_distribution(
  x,
  year_ref,
  name_pop_var = "POP",
  by = c("REGION", "SEX"),
  name_rate_var = "rate"
)
}
\arguments{
\item{x}{A database containing the variables SEX, AGE, AND REGION.}

\item{year_ref}{Reference year to use for the calculations.}

\item{name_pop_var}{Character variable with name of population variable.}

\item{by}{Character variable of combination on which the distribution should be calculated.}

\item{name_rate_var}{Name of the variable to transform in age-adjusted estimates}
}
\value{
Dataframe with population distribution and age-adjusted estimates.
}
\description{
Calculate population distribution
}
