% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{save_dw_xlsx}
\alias{save_dw_xlsx}
\title{Save the DW calculation as an xlsx-workbook}
\usage{
save_dw_xlsx(DW, folder = "", name = "", use_date = TRUE, template)
}
\arguments{
\item{DW}{result of the DW calculation (tibble)}

\item{folder}{folder, where to save the file (not needed if folder structure is integrated in name)}

\item{name}{filename to save the file (no .xlsx extension needed)}

\item{use_date}{should the date be included in the file (Default: TRUE)}

\item{template}{template that should be used when saving the file}
}
\value{
xlsx-file with on each tab the DW for each disease
}
\description{
Save the DW calculation as an xlsx-workbook
}
