% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{statbel}
\alias{statbel}
\title{STATBEL POPULATION STRUCTURE 2012-2021}
\format{
A data frame with 2222 rows and 4 variables:
\describe{
\item{SEX}{SEX - Male of Female}
\item{AGE}{AGE of the population}
\item{YEAR}{YEAR on which data was based}
\item{POP}{NUMBER of inhabitants}
...
}
}
\source{
\url{https://www.statbel.be/}
}
\usage{
statbel
}
\description{
A dataset containing the population structure from 2012 to 2021
based on the national registry.
}
\keyword{datasets}
