% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statbel_helpers.R
\name{statbel_management}
\alias{statbel_management}
\title{Perform the datamanagement of the STATBEL dataset for the BeBod-project}
\usage{
statbel_management(df_statbel, year)
}
\arguments{
\item{df_statbel}{raw STATBEL dataset that can be read in by the open.xlsx package.}

\item{year}{reference year for which the population has to be calculated.}
}
\value{
STATBEL dataset
}
\description{
Perform the datamanagement of the STATBEL dataset for the BeBod-project
}
