% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YLD.R
\name{summary1}
\alias{summary1}
\title{Generate summary}
\usage{
summary1(x)
}
\arguments{
\item{x}{Summary table from YLD (used by getYLD function)}
}
\value{
printed table from YLD calculations
}
\description{
Generate summary
}
