% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{statbel}
\alias{statbel}
\title{STATBEL POPULATION STRUCTURE 2012-2021}
\format{
A data frame with 12896 rows and 5 variables:
\describe{
\item{region}{region can be FL, BR, WA or BE}
\item{age}{age in years}
\item{gender}{gender/sex can be Men of Women}
\item{year}{year}
\item{population}{number of inhabitants in the specific strata}
...
}
}
\source{
\url{https://www.statbel.be/}
}
\usage{
statbel
}
\description{
A dataset containing the population structure from 2012 to 2021
based on the national registry.
}
\keyword{datasets}
